requirements/prd.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
Django==2.0.3
django-modelcluster==4.1
django-multiselectfield==0.1.8
django-taggit==0.22.2
django-treebeard==4.2.0
djangorestframework==3.7.7
draftjs-exporter==2.0.0
holidays==0.9.3
html5lib==0.999999999
idna==2.6
inflect==0.2.5
keyring==10.6.0
-e git://github.com/linuxsoftware/ls.joyous.git@v0.1.2#egg=ls.joyous
olefile==0.45.1
Pillow==4.3.0
pkginfo==1.4.1
psycopg2-binary==2.7.4
pycparser==2.18
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
requests-toolbelt==0.8.0
SecretStorage==2.3.1
six==1.11.0
tqdm==4.19.5
twine==1.9.1
Unidecode==0.4.21
urllib3==1.22
wagtail==2.0
webencodings==0.5.1
Willow==1.1
